% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_engine.R
\name{hermite_moments}
\alias{hermite_moments}
\title{Extract Moments from a hermite_fit Object}
\usage{
hermite_moments(object)
}
\arguments{
\item{object}{An object of class \code{"hermite_fit"}.}
}
\value{
A list with elements \code{mean}, \code{variance}, and \code{sd}.
}
\description{
Extract Moments from a hermite_fit Object
}
