% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{plot.t_hermite}
\alias{plot.t_hermite}
\title{Plot a Hermite t-Test Object}
\usage{
\method{plot}{t_hermite}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"t_hermite"}.}

\item{...}{Additional graphical parameters passed to \code{\link[graphics]{plot}}.}
}
\value{
The object \code{x}, invisibly.
}
\description{
Displays the permutation null distribution of the Hermite \eqn{t}-statistic,
marking the observed test statistic with a solid vertical line, two-sided
critical boundaries with dashed lines, and annotating the \eqn{p}-value.
}
\examples{
set.seed(42)
ctrl <- rlnorm(25, 2.0, 0.5)
trt  <- rlnorm(25, 2.3, 0.5)
res <- t_hermite(ctrl, trt, nperm = 500)
plot(res)

}
