% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarks.R
\name{ci_nct}
\alias{ci_nct}
\title{Noncentral-t (Lambda-Prime) Confidence Interval for SMD}
\usage{
ci_nct(d_point, n1, n2, conf = 0.95, df = n1 + n2 - 2L)
}
\description{
Noncentral-t (Lambda-Prime) Confidence Interval for SMD
}
