% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_hermite.R
\name{confint.cor_hermite}
\alias{confint.cor_hermite}
\title{Confidence Intervals for Hermite Correlation}
\usage{
\method{confint}{cor_hermite}(
  object,
  parm,
  level = 0.95,
  method = c("fisher", "bootstrap"),
  B = 1000L,
  ...
)
}
\description{
Confidence Intervals for Hermite Correlation
}
