% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_reg.R
\name{confint.d_reg}
\alias{confint.d_reg}
\title{Confidence Intervals for d_reg Objects}
\usage{
\method{confint}{d_reg}(
  object,
  parm,
  level = 0.95,
  method = c("bootstrap", "nct"),
  B = 1000L,
  ...
)
}
\description{
Confidence Intervals for d_reg Objects
}
