% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.d_reg}
\alias{plot.d_reg}
\title{Diagnostic Visualizations for hermiteStats Objects}
\usage{
\method{plot}{d_reg}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"d_reg"}, \code{"cor_hermite"}, or \code{"hermite_fit"}.}

\item{...}{Graphical parameters passed to base plotting functions.}
}
\description{
Diagnostic Visualizations for hermiteStats Objects
}
