% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_engine.R
\name{print.hermite_fit}
\alias{print.hermite_fit}
\title{Print Method for hermite_fit Objects}
\usage{
\method{print}{hermite_fit}(x, digits = 3L, ...)
}
\arguments{
\item{x}{An object of class \code{"hermite_fit"}.}

\item{digits}{Integer; number of decimal places to display (default = 3).}

\item{...}{Additional arguments.}
}
\description{
Print Method for hermite_fit Objects
}
