% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_engine.R
\name{summary.hermite_fit}
\alias{summary.hermite_fit}
\title{Summary Method for hermite_fit Objects}
\usage{
\method{summary}{hermite_fit}(object, digits = 3L, ...)
}
\arguments{
\item{object}{An object of class \code{"hermite_fit"}.}

\item{digits}{Integer; number of decimal places to display (default = 3).}

\item{...}{Additional arguments.}
}
\description{
Summary Method for hermite_fit Objects
}
